+++ config/module.config.php
@@ -17,7 +17,7 @@ return [
             "Core\AdapterFactory" => \Core\Adapter\AdapterFactory\AdapterFactoryFactory::class,
             "Core\BaseCacheAdapter" => \Core\Cache\CacheFactory::class,
             "Core\MemcachedAdapter" => \Core\Cache\Adapters\MemcachedAdapterFactory::class,
-            "Core\CacheAPCAdapter" => \Core\Cache\Adapters\APCAdapterFactory::class
+            "Core\CacheAPCAdapter" => \Core\Cache\Adapters\APCAdapterFactory::class,
         ]
     ]