+++ src/Core/EntityConfiguration/Error/DependencyDetailsException.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * <strong>Name :  DependencyDetailsException.php</strong></br>
+ * <strong>Desc :  [Put a description here]</strong></br>
+ *
+ * @category rdsdb-api
+ * @author Juan Pablo Cruz Maseda <pablo.cruz@digimobil.es>
+ * @copyright 2019 digimobil.es
+ * @link https://gq-api.digimobil.es
+ */
+
+
+
+class DependencyDetailsException extends \Core\Error\AbstractException
+{
+
+}