+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 
 - Implement cache with APCu, Array and Memcached/Couchebase adapters.
 - Implement Join and search by dependencies
+- Implement BaseService 
+- Implement EntityConfiguration to static class
+- Implement Filter